requirements.txt
# Core Flask dependencies
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Google API dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# OpenAI API
openai==1.3.7

# Payment processing
stripe==7.8.0

# Document processing
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pandas==2.1.3
xlrd==2.0.1

# HTTP and utilities
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Authentication and security
bcrypt==4.1.2

# Web3 dependencies for crypto payments
web3==6.11.3
eth-account==0.9.0

# Additional utilities
httpx==0.25.2
Werkzeug==3.0.1 